+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Because of a necessary database schema change, this update will deactivate all P
  - Fix bug that had the potential to lose data with certain graph display configurations
  - Prevent more than one active PID from using the same output ([#108](https://github.com/kizniche/mycodo/issues/108))
  - Prevent a PID from using the same Raise and Lower output
+ - Prevent a currently-active PID from changing the output to a currently-used output
 
 ### Miscelaneous